requirements.txt
google-generativeai==0.8.3
selenium==4.27.1
python-dotenv==1.0.1